refactor(sync): extract pairing host server into peers/host.rs
